# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rhizom API",
    "description": "REST API of a Rhizom server. Note paths are vault paths: POSIX, relative to the vault root, with extension.",
    "version": "0.0.0"
  },
...
  "paths": {
...
    "/api/graph": {
      "get": {
        "summary": "The whole vault as nodes and edges",
        "tags": ["graph"],
        "parameters": [
          {
            "schema": {
              "default": "folder",
              "anyOf": [
                {
                  "type": "string",
                  "enum": ["folder"]
                },
                {
                  "type": "string",
                  "enum": ["tag"]
                }
              ]
            },
            "in": "query",
            "name": "clusterBy",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["nodes", "edges", "clusters"],
                  "properties": {
                    "nodes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "path",
                          "name",
                          "folder",
                          "cluster",
                          "degree",
                          "inDegree",
                          "outDegree"
                        ],
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "folder": {
                            "type": "string"
                          },
                          "cluster": {
                            "type": "string"
                          },
                          "degree": {
                            "type": "integer"
                          },
                          "inDegree": {
                            "type": "integer"
                          },
                          "outDegree": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "edges": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["source", "target", "count"],
                        "properties": {
                          "source": {
                            "type": "string"
                          },
                          "target": {
                            "type": "string"
                          },
                          "count": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "clusters": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/local": {
      "get": {
        "summary": "The neighbourhood of one note",
        "tags": ["graph"],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "query",
            "name": "path",
            "required": true
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 3,
              "default": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "depth",
            "required": false
          },
          {
            "schema": {
              "default": "folder",
              "anyOf": [
                {
                  "type": "string",
                  "enum": ["folder"]
                },
                {
                  "type": "string",
                  "enum": ["tag"]
                }
              ]
            },
            "in": "query",
            "name": "clusterBy",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["nodes", "edges", "clusters"],
                  "properties": {
                    "nodes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "path",
                          "name",
                          "folder",
                          "cluster",
                          "degree",
                          "inDegree",
                          "outDegree"
                        ],
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "folder": {
                            "type": "string"
                          },
                          "cluster": {
                            "type": "string"
                          },
                          "degree": {
                            "type": "integer"
                          },
                          "inDegree": {
                            "type": "integer"
                          },
                          "outDegree": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "edges": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["source", "target", "count"],
                        "properties": {
                          "source": {
                            "type": "string"
                          },
                          "target": {
                            "type": "string"
                          },
                          "count": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "clusters": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}